+++ src/Templates/Templates/MiboMono2D/DesktopGL/Program.fs
@@ -5,7 +5,7 @@ open MiboMono2D
 
 [<EntryPoint>]
 let main _ =
-  let mgProgram = MiboMono2D.create() |> MonoGameProgram.ofProgram
+  let mgProgram = MiboMono2D.create()
 
   use game = new MiboGame<Model, Msg>(mgProgram)
   game.Run()