+++ src/Templates/Templates/MiboMono3D/WindowsDX/Program.fs
@@ -6,7 +6,7 @@ open MiboMono3D
 
 [<EntryPoint; STAThread>]
 let main _ =
-  let mgProgram = MiboMono3D.create() |> MonoGameProgram.ofProgram
+  let mgProgram = MiboMono3D.create()
 
   use game = new MiboGame<Model, Msg>(mgProgram)
   game.Run()